requirements.txt
discord.py[voice]>=2.3.2
gTTS>=2.5.1
Flask>=3.0.3
python-dotenv>=1.0.1